[HEADING] rolling back of 'displaying of codes for sub headings'
In a [PR](https://github.com/bitzesty/trade-tariff-frontend/pull/161)
we implemented displaying of codes for sub headings in a Heading details page.

But as we see now there are more cases which we need to handle in order
to display codes for sub headings properly.

So that we are going to comment in this stuff for now and return previously used
displaying of codes just for declarable commodities.

Once we finish corrects for codes for sub headings we will enable this stuff again.

[TRELLO CARD](https://trello.com/c/ND5Qd5Mc/452-tariff17-display-codes-for-headings)