Update routes to ensure that incorrect IDs result in 404s
We noticed a number of incorrect IDs being used for sections, headings,
and chapters which resulted in 500s being displayed for those users. To
avoid this and make the point of error clearer we've added constraints
to these routes meaning that if the ID doesn't match it will raise a 404
instead.