HOTFIX - Handle service choice prefix in version header and path
The `RequestForwarder` class uses the original path that was constructed
before adjustments around service choices were made in the path prefix.

This reversion to the original path is done to make sure the logs indicate the original path
that was used for the current request and because the custom
`RequestForwarder` falls out of the `ActionDispatch::Journey::Router` process around recognising a route.

The backend does not expect the prefix to be in the path or the version
so we have to remove it.

I'm leaving the `Accept` header version behaviour as is (despite it being broken) and
will sort this out later down the line.

For posterity, I believe it's behaviour should be:

1. Accept is application/vnd.uktt.v1 if /v1 is the prefix
2. Accept is application/vnd.uktt.v2 if /v2 is the prefix
3. Accept is application/vnd.uktt.v2 if the path has no version prefix