Fixes VCR failing specs that involve a request to pull chromedriver
When you execute vcr specs without chromedriver installed you'll get
test failures that show that you're missing stubbing a request to
download the latest chromedriver.

This is intermittent because some of the tests execute out of a vcr
scope and download the webdriver.

If non-vcr tests run first we have no issue (webdriver downloaded)
If vcr tests run first we have an external call to download the driver
and propagate a failure for the test